+++ src/check/index.js
@@ -1,4 +1,0 @@
-import issueCheck from './issue';
-import RedeemCheckTxParams from './redeem';
-
-export {issueCheck, RedeemCheckTxParams};