+++ src/index.js
@@ -1,12 +1,29 @@
 import {generateWallet, walletFromMnemonic, walletFromPrivateKey} from 'minterjs-wallet';
-import PostTx from './post-tx';
-import {issueCheck, RedeemCheckTxParams} from './check';
-import {SendTxParams, CreateCoinTxParams, SellTxParams, SellAllTxParams, BuyTxParams} from './coin';
-import {DeclareCandidacyTxParams, DelegateTxParams, UnbondTxParams, SetCandidateOnTxParams, SetCandidateOffTxParams} from './validator';
+import Minter from './minter';
+import MinterApi from './api';
+import PostTx from './api/post-tx';
+import EstimateCoinSell from './api/estimate-coin-sell';
+import GetNonce from './api/get-nonce';
+import issueCheck from './issue-check';
+import RedeemCheckTxParams from './tx-params/redeem-check';
+import SendTxParams from './tx-params/send';
+import CreateCoinTxParams from './tx-params/create-coin';
+import SellTxParams from './tx-params/convert-sell';
+import SellAllTxParams from './tx-params/convert-sell-all';
+import BuyTxParams from './tx-params/convert-buy';
+import DeclareCandidacyTxParams from './tx-params/candidacy-declare';
+import DelegateTxParams from '~/src/tx-params/stake-delegate';
+import UnbondTxParams from '~/src/tx-params/stake-unbond';
+import SetCandidateOnTxParams from '~/src/tx-params/candidate-set-on';
+import SetCandidateOffTxParams from '~/src/tx-params/candidate-set-off';
 
-
+export default Minter;
 export {
+    Minter,
+    MinterApi,
     PostTx,
+    EstimateCoinSell,
+    GetNonce,
     // check
     issueCheck,
     RedeemCheckTxParams,