+++ src/tx-params/convert-buy.js
@@ -1,0 +1,34 @@
+import MinterBuyTxData from 'minterjs-tx/src/tx-data/buy';
+import {formatCoin} from 'minterjs-tx/src/helpers';
+import converter from 'minterjs-tx/src/converter';
+import {TX_TYPE_BUY_COIN} from 'minterjs-tx/src/tx-types';
+
+/**
+ * @constructor
+ * @param {string} privateKey
+ * @param {string} coinFrom
+ * @param {string} coinTo
+ * @param {number} buyAmount
+ * @param {string} [feeCoinSymbol]
+ * @param {string} [message]
+ * @return {TxParams}
+ */
+export default function BuyTxParams({privateKey, coinFrom, coinTo, buyAmount, feeCoinSymbol, message}) {
+    const txData = new MinterBuyTxData({
+        coinToSell: formatCoin(coinFrom),
+        coinToBuy: formatCoin(coinTo),
+        valueToBuy: `0x${converter.convert(buyAmount, 'pip').toString(16)}`,
+    });
+
+    if (!feeCoinSymbol) {
+        feeCoinSymbol = coinFrom;
+    }
+
+    return {
+        privateKey,
+        message,
+        gasCoin: feeCoinSymbol,
+        txType: TX_TYPE_BUY_COIN,
+        txData: txData.serialize(),
+    };
+}