+++ src/variables.js
@@ -1,0 +1,2 @@
+export const API_TYPE_EXPLORER = 'explorer';
+export const API_TYPE_NODE = 'node';